fmt
docs correction

lints

fix intended logging level

fixup: manual squash step

add `helper` and the tests for it as per <https://github.com/Neptune-Crypto/neptune-core/pull/799#pullrequestreview-3993074459>
