+++ neptune-core/src/api/wallet/wallet_impl.rs
@@ -175,7 +175,6 @@ impl<'a> Wallet<'a> {
         &self,
         tx_ix: u64,
         utxo_ix: usize,
-        // block: Digest,
     ) -> Result<TxOutput, WalletError> {
         state_lock_call_async!(
             &self.state_lock,
