+++ neptune-core/src/application/mod.rs
@@ -7,5 +7,3 @@ pub mod loops;
 pub mod network;
 pub mod rpc;
 pub mod triton_vm_job_queue;
-/// auxillary proofs (not part of the protocol)
-pub mod util_proof;
